+++ apps/endringsmelding-pleiepenger/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @navikt/endringsmelding-pleiepenger
 
+## 2.25.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @navikt/sif-common-formik-ds@1.7.19
+  - @navikt/sif-common-core-ds@9.15.10
+  - @navikt/sif-common-forms-ds@8.19.7
+  - @navikt/sif-common-soknad-ds@10.13.18
+
 ## 2.25.11
 
 ### Patch Changes