+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/sif-common-ui
 
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @navikt/sif-common-formik-ds@1.7.19
+  - @navikt/sif-common-core-ds@9.15.10
+
 ## 0.6.13
 
 ### Patch Changes